+++ tests/testthat/test-check_mapbayr_data.R
@@ -27,8 +27,8 @@ test_that("DV cannot be NA if mdv == 0", {
 })
 
 test_that(".datehour columns are removed", {
-  dat <- adm_lines(cmt = 1, amt = 100, .datehour = "2022/01/01 12:34:56") %>%
-    obs_lines(cmt = 2) %>%
+  dat <- adm_rows(cmt = 1, amt = 100, .datehour = "2022/01/01 12:34:56") %>%
+    obs_rows(cmt = 2) %>%
     as.data.frame()
   expect_equal(dat[1,".datehour"], as.POSIXct("2022/01/01 12:34:56", tz = "UTC"))
   expect_null(check_mapbayr_data(dat)[[".datehour"]])