+++ tests/testthat/test-data_helpers.R
@@ -316,7 +316,7 @@ test_that("forbidden covariate works", {
 test_that("filter.mrgmod works", {
   mod <- mrgsolve::mcode("mod", "$CMT FOO", compile = FALSE)
   dat <- mod %>%
-    adm_lines(amt = c(100, 200, 300), cmt = 1) %>%
+    adm_rows(amt = c(100, 200, 300), cmt = 1) %>%
     filter(amt != 200) %>%
     get_data()
   expect_equal(dat$amt, c(100, 300))